# fix ascii.ValidPrint