I noticed an off-by-one error in the function checking whether a string contains only printable ASCII characters, where we were including 0x7F instead of excluding it.

I added a test for strings containing `~` (0x7F) and fixed the code. For some reason the fix also made the function much faster on long strings it seems:
```
name                             old time/op    new time/op    delta
ValidPrint/12345678901234567...    1.71µs ± 2%    1.24µs ± 2%  -27.14%  (p=0.000 n=9+10)

name                             old speed      new speed      delta
ValidPrint/12345678901234567...  5.86GB/s ± 2%  8.04GB/s ± 2%  +37.25%  (p=0.000 n=9+10)
```
I'll follow up with an update of the JSON benchmark to verify the impact on the code using `ascii.ValidPrint`.