@achille-roussel 2019-12-10
Short strings also show some improvement actually:
```
name                        old time/op    new time/op    delta
ValidPrint/Hello_World!       5.07ns ± 2%    4.97ns ± 1%  -2.02%  (p=0.007 n=8+8)
ValidPrint/Hello"World!       5.13ns ± 1%    4.94ns ± 1%  -3.70%  (p=0.000 n=8+8)
ValidPrint/Hello\World!       5.17ns ± 1%    4.96ns ± 1%  -4.04%  (p=0.001 n=7+7)

name                        old speed      new speed      delta
ValidPrint/Hello_World!     2.37GB/s ± 2%  2.42GB/s ± 1%  +2.07%  (p=0.010 n=8+8)
ValidPrint/Hello"World!     2.34GB/s ± 1%  2.43GB/s ± 1%  +3.84%  (p=0.000 n=8+8)
ValidPrint/Hello\World!     2.32GB/s ± 1%  2.42GB/s ± 1%  +4.24%  (p=0.001 n=7+7)
```